+++ srchybrid/kademlia/utils/ThreadName.cpp
@@ -100,7 +100,7 @@ void SetThreadName(LPCSTR szThreadName, ...)
 		THREADNAME_INFO info; 
 		info.dwType = 0x1000; 
 		info.szName = buffer; 
-		info.dwThreadID = -1; 
+		info.dwThreadID = (DWORD)-1; 
 		info.dwFlags = 0; 
 		__try 
 		{
